UIOR-244 rename variables
