+++ src/components/POLine/POLineForm.js
@@ -44,7 +44,7 @@ import getVendorsForSelect from '../Utils/getVendorsForSelect';
 import getFundsForSelect from '../Utils/getFundsForSelect';
 import getMaterialTypesForSelect from '../Utils/getMaterialTypesForSelect';
 import getIdentifierTypesForSelect from '../Utils/getIdentifierTypesForSelect';
-import getContributorsForSelect from "../Utils/getContributorsForSelect";
+import getContributorNameTypesForSelect from '../Utils/getContributorNameTypesForSelect';
 
 class POLineForm extends Component {
   static propTypes = {
@@ -212,7 +212,7 @@ class POLineForm extends Component {
     const funds = getFundsForSelect(parentResources);
     const materialTypes = getMaterialTypesForSelect(parentResources);
     const identifierTypes = getIdentifierTypesForSelect(parentResources);
-    const contributors = getContributorsForSelect(parentResources);
+    const contributorsNameTypes = getContributorNameTypesForSelect(parentResources);
     const {
       accounts,
       vendorCurrencies,
@@ -254,7 +254,7 @@ class POLineForm extends Component {
                       <ItemForm
                         formValues={formValues}
                         order={order}
-                        contributors={contributors}
+                        contributorsNameTypes={contributorsNameTypes}
                         change={change}
                         dispatch={dispatch}
                         identifierTypes={identifierTypes}
